Implement postgres storage backend

+++ internal/api/store.go
@@ -1,9 +1,9 @@
 package api
 
 type Store interface {
-	CreateSnacker(*SnackerRegistration) (uint64, error)
-	DeleteSnacker(uint64) error
+	CreateSnacker(*SnackerRegistration) (uint32, error)
+	DeleteSnacker(uint32) error
 	UpdateSnacker(*Snacker) error
-	FindSnacker(uint64) (*Snacker, error)
+	FindSnacker(uint32) (*Snacker, error)
 	FetchSnackers() ([]*Snacker, error)
 }
